pyproject.toml
[tool.poetry]
name = "common-image-tools"
version = "0.1.12"
description = ""
authors = ["Federico Lanzani <federico@federicolanzani.com>"]
readme = "README.md"
packages = [{ include = "common_image_tools" }]

[tool.poetry.dependencies]
python = "^3.10"
pillow = ">=9.5.0"
opencv-python = "^4.8.0.0"
color-transfer = "^0.1"
scikit-image = "^0.21.0"
loguru = "^0.7.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
matplotlib = "^3.7.2"
pre-commit = "^3.5.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
mypy = "^1.9.0"
ruff = "^0.4.0"


[tool.ruff]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["F841"]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
